[Chore] 로그인 화면 수정된 UI 구현
