+++ FitMate/FitMate/Resources/Assets.xcassets/logo_bgX.imageset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "logo.png",
+      "filename" : "logo_bgX.png",
       "idiom" : "universal",
       "scale" : "1x"
     },
